+++ lib/bigcommerce/prometheus.rb
@@ -22,7 +22,8 @@ require 'prometheus_exporter/server'
 require 'prometheus_exporter/client'
 require 'prometheus_exporter/middleware'
 require 'prometheus_exporter/instrumentation'
-require 'thin'
+require 'puma'
+require 'rack'
 
 require_relative 'prometheus/version'
 require_relative 'prometheus/loggable'
@@ -42,14 +43,14 @@ require_relative 'prometheus/integrations/railtie' if defined?(Rails)
 require_relative 'prometheus/integrations/puma'
 require_relative 'prometheus/integrations/resque'
 
-require_relative 'prometheus/servers/thin/server'
-require_relative 'prometheus/servers/thin/rack_app'
-require_relative 'prometheus/servers/thin/server_metrics'
-require_relative 'prometheus/servers/thin/controllers/base_controller'
-require_relative 'prometheus/servers/thin/controllers/error_controller'
-require_relative 'prometheus/servers/thin/controllers/metrics_controller'
-require_relative 'prometheus/servers/thin/controllers/not_found_controller'
-require_relative 'prometheus/servers/thin/controllers/send_metrics_controller'
+require_relative 'prometheus/servers/puma/server'
+require_relative 'prometheus/servers/puma/rack_app'
+require_relative 'prometheus/servers/puma/server_metrics'
+require_relative 'prometheus/servers/puma/controllers/base_controller'
+require_relative 'prometheus/servers/puma/controllers/error_controller'
+require_relative 'prometheus/servers/puma/controllers/metrics_controller'
+require_relative 'prometheus/servers/puma/controllers/not_found_controller'
+require_relative 'prometheus/servers/puma/controllers/send_metrics_controller'
 
 module Bigcommerce
   ##
