+++ spec/bigcommerce/prometheus/servers/puma/controllers/send_metrics_controller_spec.rb
@@ -15,8 +15,8 @@
 #
 require 'spec_helper'
 
-describe Bigcommerce::Prometheus::Servers::Thin::Controllers::MetricsController do
-  let(:request_method) { 'GET' }
+describe Bigcommerce::Prometheus::Servers::Puma::Controllers::SendMetricsController do
+  let(:request_method) { 'POST' }
   let(:env) do
     {
       'REQUEST_METHOD' => request_method,
@@ -27,40 +27,46 @@ describe Bigcommerce::Prometheus::Servers::Thin::Controllers::MetricsController 
   let(:request) { Rack::Request.new(env) }
   let(:response) { Rack::Response.new }
   let(:collector) { PrometheusExporter::Server::Collector.new }
-  let(:server_metrics) { Bigcommerce::Prometheus::Servers::Thin::ServerMetrics.new }
+  let(:server_metrics) { Bigcommerce::Prometheus::Servers::Puma::ServerMetrics.new }
   let(:controller) { described_class.new(request: request, response: response, server_metrics: server_metrics, collector: collector, logger: logger) }
-
-  let(:metrics) { ['ruby_collector_metrics_total 19', 'ruby_collector_sessions_total 19'].join("\n") }
-  let(:server_metrics_text) { ['collector_working 1', 'collector_rss 60440'].join("\n") }
-
-  before do
-    allow(collector).to receive(:prometheus_metrics_text).and_return(metrics)
-    allow(server_metrics).to receive(:to_prometheus_text).and_return(server_metrics_text)
-  end
 
   describe '#call' do
     subject { controller.call }
 
-    it 'succeeds' do
-      expect(subject.status).to eq 200
-      expect(subject.body.first).to eq 'collector_working 1
-collector_rss 60440
-ruby_collector_metrics_total 19
-ruby_collector_sessions_total 19'
-    end
+    context 'when a post request' do
+      context 'when the collector properly parses the metrics' do
+        before do
+          expect(collector).to receive(:process).once
+        end
 
-    context 'when the metrics fail to parse' do
-      let(:timeout_exception) { ::Timeout::Error }
-      before do
-        allow(collector).to receive(:prometheus_metrics_text).and_raise(timeout_exception)
+        it 'succeeds' do
+          expect(subject.status).to eq 200
+          expect(subject.body.first).to eq 'OK'
+        end
       end
 
-      it 'still responds 200, but logs an error' do
-        expect(logger).to receive(:error).with('Generating Prometheus metrics text timed out')
-        expect(subject.status).to eq 200
-        expect(subject.body.first).to eq 'collector_working 1
-collector_rss 60440
-'
+      context 'when the collector fails to parse the metrics' do
+        let(:error_message) { 'failure!' }
+        let(:exception) { StandardError.new(error_message) }
+        before do
+          expect(collector).to receive(:process).once.and_raise(exception)
+        end
+
+        it 'fails' do
+          expect(subject.status).to eq 500
+          parsed_body = JSON.parse(subject.body.first)
+          expect(parsed_body).to eq [error_message]
+        end
+      end
+    end
+
+    context 'when not a post request' do
+      let(:request_method) { 'GET' }
+
+      it 'fails' do
+        expect(subject.status).to eq 500
+        parsed_body = JSON.parse(subject.body.first)
+        expect(parsed_body).to eq ['Invalid request type. Only POST is supported.']
       end
     end
   end
